Move Link to models
