+++ src/pip/_internal/cache.py
@@ -11,6 +11,7 @@ from pip._vendor.packaging.utils import canonicalize_name
 from pip._internal import index
 from pip._internal.compat import expanduser
 from pip._internal.download import path_to_url
+from pip._internal.models.link import Link
 from pip._internal.utils.temp_dir import TempDirectory
 from pip._internal.wheel import InvalidWheelFilename, Wheel
 
@@ -101,7 +102,7 @@ class Cache(object):
         root = self.get_path_for_link(link)
         path = os.path.join(root, candidate)
 
-        return index.Link(path_to_url(path))
+        return Link(path_to_url(path))
 
     def cleanup(self):
         pass
